avoid curl ssl test if cross compiling - closes #147 - refs #148

+++ osx/scripts/build_curl.sh
@@ -64,14 +64,16 @@ $MAKE install
 mkdir -p "${BUILD}/etc/openssl/certs"
 CA_BUNDLE="${BUILD}/etc/openssl/certs/ca-bundle.crt"
 if [ ! -f ${CA_BUNDLE} ]; then
-    ${BUILD}/bin/curl --silent http://curl.haxx.se/ca/cacert.pem -o ${BUILD}/etc/openssl/certs/ca-bundle.crt
+    ${SYSTEM_CURL} --silent http://curl.haxx.se/ca/cacert.pem -o ${BUILD}/etc/openssl/certs/ca-bundle.crt
 fi
 
-# test https with cert
-echo ${BUILD}/bin/curl -I --cacert ${CA_BUNDLE} "https://www.mapbox.com/"
-${BUILD}/bin/curl -I --cacert ${CA_BUNDLE} "https://www.mapbox.com/"
-# remove curl command line now since the lack of default-known certs
-# will break other curl commands that the build system depends on
-# so we fall back to system curl command
-rm -f ${BUILD}/bin/curl
+if [[ $BOOST_ARCH != "arm" ]]; then
+    # test https with cert
+    echo ${BUILD}/bin/curl -I --cacert ${CA_BUNDLE} "https://www.mapbox.com/"
+    ${BUILD}/bin/curl -I --cacert ${CA_BUNDLE} "https://www.mapbox.com/"
+    # remove curl command line now since the lack of default-known certs
+    # will break other curl commands that the build system depends on
+    # so we fall back to system curl command
+    rm -f ${BUILD}/bin/curl
+fi
 cd ${PACKAGES}
